@@ -29,7 +29,7 @@ url = https://github.com/axiom-data-science/extract_model
 long_description = file: README.md
 long_description_content_type = text/markdown
 license = MIT
-license_file = LICENSE.txt
+license_files = LICENSE.txt
 
 ## These need to be filled in by the author!
 # For details see: https://pypi.org/classifiers/
@@ -41,9 +41,9 @@ classifiers =
     Operating System :: OS Independent
     Programming Language :: Python
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
     # Dont change this one
     License :: OSI Approved :: MIT License
 
@@ -63,7 +63,7 @@ install_requires =
 
 setup_requires=
     setuptools_scm
-python_requires = >=3.7
+python_requires = >=3.8
 ################ Up until here
 
 zip_safe = False
