@@ -4,7 +4,7 @@ Select model output using xarray and `isel` or interpolation through xarray or x
 
 import cf_xarray as cfxr  # noqa: F401
 
-from pkg_resources import DistributionNotFound, get_distribution
+from importlib.metadata import version, PackageNotFoundError
 
 import extract_model.accessor  # noqa: F401
 
@@ -13,7 +13,7 @@ from .utils import filter, order, preprocess, sub_bbox, sub_grid  # noqa: F401
 
 
 try:
-    __version__ = get_distribution("extract_model").version
-except DistributionNotFound:
+    __version__ = version("package-name")
+except PackageNotFoundError:
     # package is not installed
-    __version__ = "unknown"
+    pass
